+++ python/Makefile.am
@@ -9,13 +9,15 @@ AM_CPPFLAGS = -I$(top_builddir)/include/
 AM_CPPFLAGS += -I$(top_srcdir)/python
 AM_CPPFLAGS += @PYTHON_CFLAGS@
 
-pkgpyexec_LTLIBRARIES = _rpm.la _rpmb.la _rpms.la
+pkgpyexec_LTLIBRARIES = _rpm.la
 pkgpyexec_DATA = rpm/__init__.py rpm/transaction.py
 
 _rpm_la_LDFLAGS = -module -avoid-version -shared
 _rpm_la_LIBADD = \
 	$(top_builddir)/lib/librpm.la \
 	$(top_builddir)/rpmio/librpmio.la \
+	$(top_builddir)/build/librpmbuild.la \
+	$(top_builddir)/sign/librpmsign.la \
 	@PYTHON_LIBS@
 
 _rpm_la_SOURCES = rpmmodule.c rpmsystem-py.h \
@@ -33,24 +35,5 @@ _rpm_la_SOURCES = rpmmodule.c rpmsystem-py.h \
 	rpmstrpool-py.c rpmstrpool-py.h \
 	rpmtd-py.c rpmtd-py.h \
 	rpmte-py.c rpmte-py.h \
-	rpmts-py.c rpmts-py.h
-
-_rpmb_la_LDFLAGS = -module -avoid-version -shared
-_rpmb_la_LIBADD = \
-	$(top_builddir)/build/librpmbuild.la \
-	$(top_builddir)/lib/librpm.la \
-	$(top_builddir)/rpmio/librpmio.la \
-	@PYTHON_LIBS@
-
-_rpmb_la_SOURCES = rpmbmodule.c rpmsystem-py.h \
+	rpmts-py.c rpmts-py.h \
 	spec-py.c spec-py.h
-
-_rpms_la_LDFLAGS = -module -avoid-version -shared
-_rpms_la_LIBADD = \
-        $(top_builddir)/sign/librpmsign.la \
-        $(top_builddir)/lib/librpm.la \
-        $(top_builddir)/rpmio/librpmio.la \
-	@PYTHON_LIBS@
-
-_rpms_la_SOURCES = rpmsmodule.c rpmsystem-py.h
-
