+++ python/rpm/__init__.py
@@ -46,18 +46,6 @@ _RPMVSF_NOSIGNATURES = _rpm._RPMVSF_NOSIGNATURES
 __version__ = _rpm.__version__
 __version_info__ = tuple(__version__.split('.'))
 
-# try to import build bits but dont require it
-try:
-    from rpm._rpmb import *
-except ImportError:
-    pass
-
-# try to import signing bits but dont require it
-try:
-    from rpm._rpms import *
-except ImportError:
-    pass
-
 # backwards compatibility + give the same class both ways
 ts = TransactionSet
