+++ python/setup.py.in
@@ -24,36 +24,18 @@ if os.access('Makefile.am', os.F_OK):
                                  '-Wl,-L../sign/.libs'])
     os.environ['PKG_CONFIG_PATH'] = '..'
 
-
 rpmmod = Extension('rpm._rpm',
-                   sources = [  'header-py.c', 'rpmds-py.c', 'rpmfd-py.c',
-				'rpmfi-py.c', 'rpmii-py.c', 'rpmkeyring-py.c',
-                                'rpmmacro-py.c', 'rpmmi-py.c', 'rpmps-py.c',
-                                'rpmstrpool-py.c', 'rpmfiles-py.c', 
-				'rpmarchive-py.c', 'rpmtd-py.c',
-                                'rpmte-py.c', 'rpmts-py.c', 'rpmmodule.c',
-                             ],
+                   sources = ['header-py.c', 'rpmds-py.c', 'rpmfd-py.c',
+                              'rpmfi-py.c', 'rpmii-py.c', 'rpmkeyring-py.c',
+                              'rpmmacro-py.c', 'rpmmi-py.c', 'rpmps-py.c',
+                              'rpmstrpool-py.c', 'rpmfiles-py.c',
+                              'rpmarchive-py.c', 'rpmtd-py.c',
+                              'rpmte-py.c', 'rpmts-py.c',
+                              'spec-py',
+                              'rpmmodule.c'],
                    include_dirs = pkgconfig('--cflags'),
                    library_dirs = pkgconfig('--libs-only-L'),
-                   libraries = pkgconfig('--libs-only-l'),
-                   extra_compile_args = cflags,
-                   extra_link_args = additional_link_args
-                  )
-
-rpmbuild_mod = Extension('rpm._rpmb',
-                   sources = ['rpmbmodule.c', 'spec-py.c'],
-                   include_dirs = pkgconfig('--cflags'),
-                   library_dirs = pkgconfig('--libs-only-L'),
-                   libraries = pkgconfig('--libs-only-l') + ['rpmbuild'],
-                   extra_compile_args = cflags,
-                   extra_link_args = additional_link_args
-                  )
-
-rpmsign_mod = Extension('rpm._rpms',
-                   sources = ['rpmsmodule.c'],
-                   include_dirs = pkgconfig('--cflags'),
-                   library_dirs = pkgconfig('--libs-only-L'),
-                   libraries = pkgconfig('--libs-only-l') + ['rpmsign'],
+                   libraries = pkgconfig('--libs-only-l') + ['rpmbuild', 'rpmsign'],
                    extra_compile_args = cflags,
                    extra_link_args = additional_link_args
                   )
@@ -63,6 +45,6 @@ setup(name='@PACKAGE_NAME@',
       description='Python bindings for @PACKAGE_NAME@',
       maintainer_email='@PACKAGE_BUGREPORT@',
       url='http://www.rpm.org/',
-      packages = ['@PACKAGE_NAME@'],
-      ext_modules= [rpmmod, rpmbuild_mod, rpmsign_mod]
+      packages=['@PACKAGE_NAME@'],
+      ext_modules=[rpmmod]
      )
